+++ contracts/registry/src/state.rs
@@ -8,7 +8,7 @@ use cw_storage_plus::Map;
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct CONTRACT {
     pub gas_limit: u64,
-    pub gas_price: u128,
+    pub gas_price: u64,
     pub is_executable: bool,
 }
